refactor: route delay view with separate modules for sub-views
