+++ frontend/src/views/routeDelayView.tsx
@@ -1,25 +1,14 @@
-import FormControl from "@mui/material/FormControl";
-import InputLabel from "@mui/material/InputLabel";
-import MenuItem from "@mui/material/MenuItem";
-import Pagination from "@mui/material/Pagination";
-import Select from "@mui/material/Select";
 import ToggleButton from "@mui/material/ToggleButton";
 import ToggleButtonGroup from "@mui/material/ToggleButtonGroup";
-import ArrowForwardIosIcon from "@mui/icons-material/ArrowForwardIos";
-import type { SelectChangeEvent } from "@mui/material/Select";
-import { DelaySummary } from "../types/historicalDelay";
 import { RouteDelayControls } from "../components/RouteDelayControls";
-import { RouteDetailsPage } from "../components/RouteDetailsPage";
 import { DatePreset, EventType, getPresetDescription } from "../types/departureDelay";
 import { TransportationMode } from "../types/sl";
-import {
-    PageSizeOption,
-    PageSizeOptions,
-    RouteDelaySection,
-    RouteDelayTrendPoint,
-} from "../types/routeDelays";
-import { getAvgDelayMinutes, getDelayTextColorClass } from "../utils/time";
-import { getRouteDisplayName, getRouteIdentityKey } from "../utils/route";
+import { PageSizeOption, RouteDelaySection, RouteDelayTrendPoint } from "../types/routeDelays";
+import { DelaySummary } from "../types/historicalDelay";
+import { RouteDelayRoutesView } from "./routeDelayRoutesView";
+import { RouteDelayLeaderboardView } from "./routeDelayLeaderboardView";
+import { RouteDelayRouteFallbackView } from "./routeDelayRouteFallbackView";
+import { RouteDetailsPage } from "../components/RouteDetailsPage";
 
 type RouteDelayViewProps = {
     selectedSection: RouteDelaySection;
@@ -33,7 +22,7 @@ type RouteDelayViewProps = {
     currentPage: number;
     totalPages: number;
     totalFilteredRoutes: number;
-    routesPerPage: number;
+    routesPerPage: PageSizeOption;
     selectedRouteKey: string | null;
     selectedRouteSummary: DelaySummary | null;
     leaderboardItems: DelaySummary[];
@@ -48,6 +37,7 @@ type RouteDelayViewProps = {
     onSearchQueryChange: (query: string) => void;
     onSelectedSectionChange: (section: RouteDelaySection) => void;
     onSelectRoute: (routeKey: string | null) => void;
+    onBackToRoutes: () => void;
     onPageChange: (nextPage: number) => void;
     onRoutesPerPageChange: (nextPageSize: PageSizeOption) => void;
 };
@@ -79,6 +69,7 @@ export function RouteDelayView({
     onSearchQueryChange,
     onSelectedSectionChange,
     onSelectRoute,
+    onBackToRoutes,
     onPageChange,
     onRoutesPerPageChange,
 }: RouteDelayViewProps) {
@@ -90,78 +81,6 @@ export function RouteDelayView({
         selectedSection === "routes"
             ? `Showing ${pagedRoutes.length} of ${totalFilteredRoutes} filtered routes`
             : `Showing ${totalFilteredRoutes} filtered routes`;
-
-    function getRouteListItemCB(summary: DelaySummary) {
-        const routeKey = getRouteIdentityKey(summary);
-        const avgDelayMinutes = getAvgDelayMinutes(summary, selectedEventType);
-
-        function handleClickACB() {
-            onSelectRoute(routeKey);
-        }
-
-        return (
-            <li key={routeKey}>
-                <button
-                    type="button"
-                    className={`w-full rounded border p-2 text-left transition 
-                        border-slate-200 hover:border-slate-300 hover:bg-slate-50
-                        flex items-center justify-between`}
-                    onClick={handleClickACB}
-                >
-                    <div>
-                        <p className="text-sm font-semibold text-slate-900">
-                            {summary.route?.shortName} {summary.route?.longName}
-                        </p>
-                        <p className="text-xs text-slate-600">
-                            Average delay:{" "}
-                            <span className={getDelayTextColorClass(avgDelayMinutes)}>
-                                {avgDelayMinutes} min
-                            </span>
-                            , {summary.uniqueTrips} unique trips
-                        </p>
-                    </div>
-                    <ArrowForwardIosIcon className="mr-2" />
-                </button>
-            </li>
-        );
-    }
-
-    function getLeaderboardItemCB(summary: DelaySummary, index: number) {
-        const avgDelayMinutes = getAvgDelayMinutes(summary, selectedEventType);
-        return (
-            <li
-                key={getRouteIdentityKey(summary)}
-                className="flex items-center justify-between text-sm text-slate-700"
-            >
-                <span className="flex items-center gap-3">
-                    <span className="w-6 text-right font-semibold text-slate-500 tabular-nums">
-                        {index + 1}.
-                    </span>
-                    <span>{getRouteDisplayName(summary)}</span>
-                </span>
-                <span className="flex font-medium tabular-nums">
-                    <span className="text-right">{avgDelayMinutes} min</span>
-                    <span className="w-36 text-right">{summary.uniqueTrips} unique trips</span>
-                </span>
-            </li>
-        );
-    }
-
-    function getRoutesPerPageOptionCB(pageSize: number) {
-        return (
-            <MenuItem key={pageSize} value={pageSize}>
-                {pageSize} / page
-            </MenuItem>
-        );
-    }
-
-    function handleRoutesPerPageChangeACB(event: SelectChangeEvent<number>) {
-        onRoutesPerPageChange(event.target.value as PageSizeOption);
-    }
-
-    function handlePaginationChangeACB(_: React.ChangeEvent<unknown>, nextPage: number) {
-        onPageChange(nextPage);
-    }
 
     function handleSectionChangeACB(
         _: React.MouseEvent<HTMLElement>,
@@ -219,60 +138,23 @@ export function RouteDelayView({
                         {selectedRouteKey === null ? (
                             <div className="flex flex-col gap-4 pt-4">
                                 {selectedSection === "routes" ? (
-                                    <section className="space-y-3">
-                                        <div className="flex flex-wrap items-center justify-between gap-3">
-                                            <FormControl size="small" sx={{ minWidth: 125 }}>
-                                                <InputLabel id="routes-per-page-label">
-                                                    Rows
-                                                </InputLabel>
-                                                <Select
-                                                    labelId="routes-per-page-label"
-                                                    label="Rows"
-                                                    value={routesPerPage}
-                                                    onChange={handleRoutesPerPageChangeACB}
-                                                >
-                                                    {PageSizeOptions.map(getRoutesPerPageOptionCB)}
-                                                </Select>
-                                            </FormControl>
-
-                                            {totalPages > 1 ? (
-                                                <Pagination
-                                                    count={totalPages}
-                                                    page={currentPage}
-                                                    onChange={handlePaginationChangeACB}
-                                                    color="primary"
-                                                    shape="rounded"
-                                                    size="small"
-                                                    siblingCount={1}
-                                                    boundaryCount={1}
-                                                />
-                                            ) : null}
-                                        </div>
-
-                                        {pagedRoutes.length === 0 ? (
-                                            <p className="rounded border border-slate-200 p-3 text-sm text-slate-500">
-                                                No routes match the selected filters.
-                                            </p>
-                                        ) : (
-                                            <ul className="space-y-2">
-                                                {pagedRoutes.map(getRouteListItemCB)}
-                                            </ul>
-                                        )}
-                                    </section>
+                                    <RouteDelayRoutesView
+                                        pagedRoutes={pagedRoutes}
+                                        currentPage={currentPage}
+                                        totalPages={totalPages}
+                                        routesPerPage={routesPerPage}
+                                        selectedEventType={selectedEventType}
+                                        onSelectRoute={onSelectRoute}
+                                        onPageChange={onPageChange}
+                                        onRoutesPerPageChange={onRoutesPerPageChange}
+                                    />
                                 ) : null}
 
                                 {selectedSection === "leaderboard" ? (
-                                    <section className="space-y-2">
-                                        {leaderboardItems.length === 0 ? (
-                                            <p className="rounded border border-slate-200 p-3 text-sm text-slate-500">
-                                                No leaderboard data available.
-                                            </p>
-                                        ) : (
-                                            <ol className="space-y-1 rounded border border-slate-200 p-3">
-                                                {leaderboardItems.map(getLeaderboardItemCB)}
-                                            </ol>
-                                        )}
-                                    </section>
+                                    <RouteDelayLeaderboardView
+                                        leaderboardItems={leaderboardItems}
+                                        selectedEventType={selectedEventType}
+                                    />
                                 ) : null}
                             </div>
                         ) : selectedRouteSummary ? (
@@ -281,22 +163,10 @@ export function RouteDelayView({
                                 selectedEventType={selectedEventType}
                                 trendPoints={trendPoints}
                                 isTrendLoading={isTrendLoading}
-                                onBack={() => onSelectRoute(null)}
+                                onBackToRoutes={onBackToRoutes}
                             />
                         ) : (
-                            <div className="space-y-3 rounded border border-slate-200 p-4 text-sm text-slate-700">
-                                <p>
-                                    The selected route is no longer available for the current
-                                    filters.
-                                </p>
-                                <button
-                                    type="button"
-                                    className="rounded border border-slate-300 p-2 font-medium transition hover:bg-slate-50"
-                                    onClick={() => onSelectRoute(null)}
-                                >
-                                    Back to routes
-                                </button>
-                            </div>
+                            <RouteDelayRouteFallbackView onBackToRoutes={onBackToRoutes} />
                         )}
                     </div>
                 </div>
